base: modify the mundane
Monet is good, but we want to go crazy sometimes.
That's why we're trying to mix monet with
a bit of transparency.

It's a really ugly attempt and it's not fully supported.
It would be cool to have a wider monet transparency
compatibility but until now, let's add a resource
on our system to generate a transparent variant
of the wallpaper color.

Taking notes from valerie:
- https://play-lh.googleusercontent.com/FFXP6eaNbBNIJGhUelE1V0TNQ8bYypOIKcys_qwjaWP1YHHZat9PE9xZkE_WDEPQEa-i=w2560-h1440-rw
and calypso:
- https://raw.githubusercontent.com/aleoponcelet/substratum-calypso-theme/main/Screenshots/Fi1.png

TODO: Implement a more controllable version of this
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>